+++ wuye/wuye/wuye/Utilities.m
@@ -128,7 +128,7 @@ NSInteger minorVersion = -1;
 {
     CC_MD5_CTX ctx;
     CC_MD5_Init(&ctx);
-    CC_MD5_Update(&ctx, buffer, length);
+    CC_MD5_Update(&ctx, buffer, (unsigned int)length);
     char *md;
     CC_MD5_Final((unsigned char *)md, &ctx);
     NSString *ret;
